main landed P3-07, P3-08 (column inspector) and P3-11 (drawer tabs). One
conflict, in the inspector: main rewrote it to read an `inspector` component
theme instead of the sheet, so its side wins outright and this branch's
mechanical `colors` -> `colors()` rename there is moot.

The inspector's theme entries were authored against the old model, though, so
they arrived duplicating tokens this branch names. Seven fields whose value
names the same slot in *both* themes are collapsed to references
(label_color, meta_color, note_color, divider_fill, null_color,
format_parquet_color, format_view_color); both themes still resolve
identically. format_json_color and format_csv_color disagree across themes,
so they stay specific alongside the other cross-theme divergences already
noted for the designer pass.

+++ crates/strata-freya/src/apps/project/views/workbench/results/record_view.rs
@@ -26,11 +26,11 @@ use strata_model::Kind;
 use super::cell_view::page_batch_row;
 use super::copy;
 use super::datagrid::GridData;
-use super::status_bar::fmt_int;
 use crate::components::divider::Divider;
 use crate::components::icon::{Icon, IconName};
 use crate::components::type_palette::{kind_color, type_palette};
 use crate::components::typography::{Meta, MonoValue, Path, Readout};
+use strata_core::util::fmt_int;
 
 define_theme!(
     %[component]
@@ -177,8 +177,12 @@ impl Component for RecordView {
             .spacing(8.)
             .padding((12., 16.))
             .child(
-                MonoValue::new(format!("Row {} of {}", fmt_int(abs_n), fmt_int(self.total)))
-                    .color(theme.label_color),
+                MonoValue::new(format!(
+                    "Row {} of {}",
+                    fmt_int(abs_n as u64),
+                    fmt_int(self.total as u64)
+                ))
+                .color(theme.label_color),
             )
             .child(rect().width(Size::flex(1.)))
             .child(copy_button("JSON", "Copy row as JSON", {
